+++ src/main/java/com/umc/linkyou/batch/curation/GenerateMonthlyCurationJobConfig.java
@@ -12,8 +12,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.transaction.PlatformTransactionManager;
 
 /**
- * 매월 1일마다 유저별 큐레이션을 생성하고, 이어서 생성된 큐레이션에 대한 알림을 발송하는 배치 작업 설정
- * 두 Step 모두 청크 기반으로 구현
+ * 매월 1일마다 유저별 큐레이션을 생성하는 배치 작업 설정
+ * 생성된 큐레이션에 대한 알림 발송은 발송 시각이 달라(오전 8시) {@link SendCurationAlarmJobConfig}의 별도 Job으로 분리되어 있다.
  */
 @Configuration
 @RequiredArgsConstructor
@@ -26,15 +26,11 @@ public class GenerateMonthlyCurationJobConfig {
     private final CurationItemReader curationItemReader;
     private final CurationItemProcessor curationItemProcessor;
     private final CurationItemWriter curationItemWriter;
-    private final CurationAlarmItemReader curationAlarmItemReader;
-    private final CurationAlarmItemProcessor curationAlarmItemProcessor;
-    private final CurationAlarmItemWriter curationAlarmItemWriter;
 
     @Bean
     public Job generateMonthlyCurationJob() {
         return new JobBuilder("generateMonthlyCurationJob", jobRepository)
                 .start(generateMonthlyCurationStep())
-                .next(sendCurationAlarmStep())
                 .build();
     }
 
@@ -45,18 +41,6 @@ public class GenerateMonthlyCurationJobConfig {
                 .reader(curationItemReader)
                 .processor(curationItemProcessor)
                 .writer(curationItemWriter)
-                .build();
-    }
-
-    // 큐레이션 생성 실패와 알림 발송 실패를 분리하기 위해 별도 Step으로 구성한다.
-    // 발송 대상은 Alarm 존재 여부로 판단하므로, 이 Step만 재실행돼도 이미 보낸 유저는 다시 대상이 되지 않는다.
-    @Bean
-    public Step sendCurationAlarmStep() {
-        return new StepBuilder("sendCurationAlarmStep", jobRepository)
-                .<CurationAlarmCandidate, CurationAlarmBatchItem>chunk(CHUNK_SIZE, transactionManager)
-                .reader(curationAlarmItemReader)
-                .processor(curationAlarmItemProcessor)
-                .writer(curationAlarmItemWriter)
                 .build();
     }
 }
